backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
python-dotenv==1.1.1
PyJWT[crypto]==2.10.1
pydantic==2.11.7
python-jose[cryptography]
pinecone==8.1.2
psycopg2-binary==2.9.11
requests==2.33.1
